:bug: profile 수정 api favTopics 아이디만 주게끔 변경 (#125)

+++ src/main/java/com/hsu/mamomo/repository/jpa/FavTopicRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FavTopicRepository extends JpaRepository<FavTopic, Long> {
+public interface FavTopicRepository extends JpaRepository<FavTopic, Integer> {
 
     Optional<List<FavTopic>> findFavTopicByUser(User user);
 }
